// client/rest/models/cookup-jsonld.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Cookup
 * @export
 * @interface CookupJsonld
 */
export interface CookupJsonld {
    /**
     * 
     * @type {string | any}
     * @memberof CookupJsonld
     */
    '@context'?: string | any;
    /**
     * 
     * @type {string}
     * @memberof CookupJsonld
     */
    '@id'?: string;
    /**
     * 
     * @type {string}
     * @memberof CookupJsonld
     */
    '@type'?: string;
    /**
     * 
     * @type {number}
     * @memberof CookupJsonld
     */
    'cookupId'?: number;
    /**
     * 
     * @type {string}
     * @memberof CookupJsonld
     */
    'user'?: string;
    /**
     * 
     * @type {string}
     * @memberof CookupJsonld
     */
    'recipe'?: string;
    /**
     * 
     * @type {string}
     * @memberof CookupJsonld
     */
    'date'?: string;
}